backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
python-jose[cryptography]==3.4.0
pydantic-settings==2.11.0
python-dotenv==1.1.1
python-multipart==0.0.20
supabase==2.18.1
httpx==0.28.1
beautifulsoup4==4.13.4
selenium==4.35.0
pillow==11.3.0
pytesseract==0.3.13
